feat: cart, coupons, products 별 entities 분리
